requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.10.0
cycler==0.12.1
fonttools==4.57.0
frozendict==2.4.6
hdbscan==0.8.40
idna==3.10
joblib==1.5.0
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.4.0
matplotlib==3.10.1
multitasking==0.0.11
narwhals==1.37.1
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pycparser==2.22
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
ta==0.11.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
yfinance==0.2.58
openai>=1.0.0
feedparser
python-dotenv
kaleido 
openpyxl>=3.1.0
tqdm>=4.64.1
fastapi
uvicorn
pytz
python-multipart
orjson
apscheduler
yahooquery
finvizfinance
transformers>=4.38.0
torch>=2.2.0
pytrends
textblob
playwright
sentencepiece
fake-useragent
